requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
filelock==3.12.0
Flask==2.2.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
fsspec==2023.5.0
greenlet==2.0.2
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
packaging==23.1
Pillow==9.5.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
python-dotenv==1.0.0
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
SQLAlchemy==2.0.9
sympy==1.12
tokenizers==0.13.3
torch==2.0.1
tqdm==4.65.0
transformers==4.29.0
typing_extensions==4.5.0
urllib3==2.0.2
Werkzeug==2.2.3
zipp==3.15.0
